src/bot/requirements.txt
kink
discord.py
python-dotenv
PyNaCl
pytubefix
pydantic